01_day/uv_package/pyproject.toml
[project]
name = "nescom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sir Qasim", email = "m.qasim077@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.30.7",
    "crewai>=0.19.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0"
]

[project.scripts]
nescom = "nescom:main"
abc = "nescom.hello:my_function"
hello = "nescom.hello:say_hello"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
